requirements.txt
matplotlib
requests
pandas
pymongo
python-dotenv
fetch_data